+++ samples/cellular/modem_shell/src/sock/sock.c
@@ -686,7 +686,7 @@ static void print_throughput_summary(uint32_t data_len, int64_t time_ms)
 
 	mosh_print("Summary:");
 	mosh_print("  Data length: %7u bytes", data_len);
-	mosh_print("  Time:       %8.3f s", (float)time_ms / 1000);
+	mosh_print("  Time:       %8.3f s", (double)time_ms / 1000.0);
 	mosh_print("  Throughput:  %7.0f bit/s", throughput);
 }
 
@@ -836,7 +836,7 @@ static void sock_send_random_data_length(struct sock_info *socket_info)
 			mosh_print(
 				"%7u bytes, %6.2fs, %6.0f bit/s",
 				socket_info->send_bytes_sent,
-				(float)ul_time_intermediate_ms / 1000,
+				(double)ul_time_intermediate_ms / 1000.0,
 				throughput);
 
 			socket_info->send_print_interval += 10;
